[participant-integration-api] - Split StandaloneIndexerServer into IndexerServer and IndexerServerMigrations to allow us to run them in separate order, so we can first run migrations and then instantiate other services which depend on the migrations (JdbcIndexer)
changelog_begin
changelog_end
